requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
python-dotenv
chainlit
openai
